+++ ios/RCTMapboxGL/RCTMapboxGLManager.m
@@ -15,7 +15,6 @@
 #import "RCTEventDispatcher.h"
 #import "UIView+React.h"
 #import "RCTUIManager.h"
-#import "RCTSparseArray.h"
 
 @implementation RCTMapboxGLManager
 
@@ -39,9 +38,33 @@ RCT_EXPORT_MODULE();
       @"onRegionWillChange",
       @"onOpenAnnotation",
       @"onRightAnnotationTapped",
-      @"onUpdateUserLocation"
+      @"onUpdateUserLocation",
+      @"onTap",
+      @"onLongPress"
     ];
 }
+
+- (NSDictionary *)constantsToExport
+{
+    return @{
+             @"mapStyles": @{
+                 @"light": [[MGLStyle lightStyleURL] absoluteString],
+                 @"dark": [[MGLStyle darkStyleURL] absoluteString],
+                 @"streets": [[MGLStyle streetsStyleURL] absoluteString],
+                 @"emerald": [[MGLStyle emeraldStyleURL] absoluteString],
+                 @"satellite": [[MGLStyle satelliteStyleURL] absoluteString],
+                 @"hybrid": [[MGLStyle hybridStyleURL] absoluteString],
+            },
+             @"userTrackingMode": @{
+                     @"none": [NSNumber numberWithUnsignedInt:MGLUserTrackingModeNone],
+                     @"follow": [NSNumber numberWithUnsignedInt:MGLUserTrackingModeFollow],
+                     @"followWithCourse": [NSNumber numberWithUnsignedInt:MGLUserTrackingModeFollowWithCourse],
+                     @"followWithHeading": [NSNumber numberWithUnsignedInt:MGLUserTrackingModeFollowWithHeading]
+            }
+    };
+
+};
+
 
 RCT_EXPORT_VIEW_PROPERTY(accessToken, NSString);
 RCT_EXPORT_VIEW_PROPERTY(centerCoordinate, CLLocationCoordinate2D);
@@ -53,11 +76,31 @@ RCT_EXPORT_VIEW_PROPERTY(scrollEnabled, BOOL);
 RCT_EXPORT_VIEW_PROPERTY(zoomEnabled, BOOL);
 RCT_EXPORT_VIEW_PROPERTY(showsUserLocation, BOOL);
 RCT_EXPORT_VIEW_PROPERTY(styleURL, NSURL);
+RCT_EXPORT_VIEW_PROPERTY(userTrackingMode, int);
 RCT_EXPORT_VIEW_PROPERTY(zoomLevel, double);
+
+RCT_CUSTOM_VIEW_PROPERTY(attributionButtonIsHidden, BOOL, RCTMapboxGL)
+{
+    BOOL value = [json boolValue];
+    [view setAttributionButtonVisibility:value ? true : false];
+}
+
+RCT_CUSTOM_VIEW_PROPERTY(logoIsHidden, BOOL, RCTMapboxGL)
+{
+    BOOL value = [json boolValue];
+    [view setLogoVisibility:value ? true : false];
+}
+
+RCT_CUSTOM_VIEW_PROPERTY(compassIsHidden, BOOL, RCTMapboxGL)
+{
+    BOOL value = [json boolValue];
+    [view setCompassVisibility:value ? true : false];
+}
+
 RCT_EXPORT_METHOD(setZoomLevelAnimated:(nonnull NSNumber *)reactTag
                   zoomLevel:(double)zoomLevel)
 {
-    [_bridge.uiManager addUIBlock:^(RCTUIManager *uiManager, RCTSparseArray *viewRegistry) {
+    [_bridge.uiManager addUIBlock:^(RCTUIManager *uiManager, NSDictionary<NSNumber *, UIView *> *viewRegistry) {
         RCTMapboxGL *mapView = viewRegistry[reactTag];
         if ([mapView isKindOfClass:[RCTMapboxGL class]]) {
             [mapView setZoomLevelAnimated:zoomLevel];
@@ -67,7 +110,7 @@ RCT_EXPORT_METHOD(setZoomLevelAnimated:(nonnull NSNumber *)reactTag
 RCT_EXPORT_METHOD(setDirectionAnimated:(nonnull NSNumber *)reactTag
                   heading:(float)heading)
 {
-    [_bridge.uiManager addUIBlock:^(RCTUIManager *uiManager, RCTSparseArray *viewRegistry) {
+    [_bridge.uiManager addUIBlock:^(RCTUIManager *uiManager, NSDictionary<NSNumber *, UIView *> *viewRegistry) {
         RCTMapboxGL *mapView = viewRegistry[reactTag];
         if ([mapView isKindOfClass:[RCTMapboxGL class]]) {
             [mapView setDirectionAnimated:heading];
@@ -79,7 +122,7 @@ RCT_EXPORT_METHOD(setCenterCoordinateAnimated:(nonnull NSNumber *)reactTag
                   latitude:(float) latitude
                   longitude:(float) longitude)
 {
-    [_bridge.uiManager addUIBlock:^(RCTUIManager *uiManager, RCTSparseArray *viewRegistry) {
+    [_bridge.uiManager addUIBlock:^(RCTUIManager *uiManager, NSDictionary<NSNumber *, UIView *> *viewRegistry) {
         RCTMapboxGL *mapView = viewRegistry[reactTag];
         if ([mapView isKindOfClass:[RCTMapboxGL class]]) {
             [mapView setCenterCoordinateAnimated:CLLocationCoordinate2DMake(latitude, longitude)];
@@ -92,7 +135,7 @@ RCT_EXPORT_METHOD(setCenterCoordinateZoomLevelAnimated:(nonnull NSNumber *)react
                   longitude:(float) longitude
                   zoomLevel:(double)zoomLevel)
 {
-    [_bridge.uiManager addUIBlock:^(RCTUIManager *uiManager, RCTSparseArray *viewRegistry) {
+    [_bridge.uiManager addUIBlock:^(RCTUIManager *uiManager, NSDictionary<NSNumber *, UIView *> *viewRegistry) {
         RCTMapboxGL *mapView = viewRegistry[reactTag];
         if ([mapView isKindOfClass:[RCTMapboxGL class]]) {
             [mapView setCenterCoordinateZoomLevelAnimated:CLLocationCoordinate2DMake(latitude, longitude) zoomLevel:zoomLevel];
@@ -107,7 +150,7 @@ RCT_EXPORT_METHOD(setVisibleCoordinateBoundsAnimated:(nonnull NSNumber *)reactTa
                   longitudeNE:(float) longitudeNE
                   edgePadding:(double) edgePadding)
 {
-    [_bridge.uiManager addUIBlock:^(RCTUIManager *uiManager, RCTSparseArray *viewRegistry) {
+    [_bridge.uiManager addUIBlock:^(RCTUIManager *uiManager, NSDictionary<NSNumber *, UIView *> *viewRegistry) {
         RCTMapboxGL *mapView = viewRegistry[reactTag];
         if ([mapView isKindOfClass:[RCTMapboxGL class]]) {
             MGLCoordinateBounds coordinatesBounds = MGLCoordinateBoundsMake(CLLocationCoordinate2DMake(latitudeSW, longitudeSW), CLLocationCoordinate2DMake(latitudeNE, longitudeNE));
@@ -119,7 +162,7 @@ RCT_EXPORT_METHOD(setVisibleCoordinateBoundsAnimated:(nonnull NSNumber *)reactTa
 RCT_EXPORT_METHOD(selectAnnotationAnimated:(nonnull NSNumber *) reactTag
                   annotationInArray:(NSUInteger)annotationInArray)
 {
-    [_bridge.uiManager addUIBlock:^(RCTUIManager *uiManager, RCTSparseArray *viewRegistry) {
+    [_bridge.uiManager addUIBlock:^(RCTUIManager *uiManager, NSDictionary<NSNumber *, UIView *> *viewRegistry) {
         RCTMapboxGL *mapView = viewRegistry[reactTag];
         if ([mapView isKindOfClass:[RCTMapboxGL class]]) {
             [mapView selectAnnotationAnimated:annotationInArray];
@@ -130,7 +173,7 @@ RCT_EXPORT_METHOD(selectAnnotationAnimated:(nonnull NSNumber *) reactTag
 RCT_EXPORT_METHOD(removeAnnotation:(nonnull NSNumber *) reactTag
                   annotationInArray:(NSUInteger)annotationInArray)
 {
-    [_bridge.uiManager addUIBlock:^(RCTUIManager *uiManager, RCTSparseArray *viewRegistry) {
+    [_bridge.uiManager addUIBlock:^(RCTUIManager *uiManager, NSDictionary<NSNumber *, UIView *> *viewRegistry) {
         RCTMapboxGL *mapView = viewRegistry[reactTag];
         if ([mapView isKindOfClass:[RCTMapboxGL class]]) {
             [mapView removeAnnotation:annotationInArray];
@@ -138,13 +181,23 @@ RCT_EXPORT_METHOD(removeAnnotation:(nonnull NSNumber *) reactTag
     }];
 }
 
+RCT_EXPORT_METHOD(setUserTrackingMode:(nonnull NSNumber *) reactTag
+                  userTrackingMode:(int)userTrackingMode)
+{
+    [_bridge.uiManager addUIBlock:^(RCTUIManager *uiManager, NSDictionary<NSNumber *, UIView *> *viewRegistry) {
+        RCTMapboxGL *mapView = viewRegistry[reactTag];
+        if ([mapView isKindOfClass:[RCTMapboxGL class]]) {
+            [mapView setUserTrackingMode:userTrackingMode];
+        }
+    }];
+}
+
 RCT_EXPORT_METHOD(addAnnotations:(nonnull NSNumber *)reactTag
                   annotations:(NSArray *) annotations)
 {
-    [_bridge.uiManager addUIBlock:^(RCTUIManager *uiManager, RCTSparseArray *viewRegistry) {
+    [_bridge.uiManager addUIBlock:^(RCTUIManager *uiManager, NSDictionary<NSNumber *, UIView *> *viewRegistry) {
         RCTMapboxGL *mapView = viewRegistry[reactTag];
         if([mapView isKindOfClass:[RCTMapboxGL class]]) {
-            
             NSMutableArray* annotationsArray = [NSMutableArray array];
             id annotationObject;
             NSEnumerator *enumerator = [annotations objectEnumerator];