+++ ios/example.js
@@ -79,6 +79,9 @@ var MapExample = React.createClass({
   onRightAnnotationTapped(e) {
     console.log(e);
   },
+  onLongPress(location) {
+    console.log('long pressed', location);
+  },
   render: function() {
     StatusBarIOS.setHidden(true);
     return (
@@ -119,6 +122,9 @@ var MapExample = React.createClass({
         <Text style={styles.text} onPress={() => this.setVisibleCoordinateBoundsAnimated(mapRef, 40.712, -74.227, 40.774, -74.125, 0)}>
           Set visible bounds to 40.7, -74.2, 40.7, -74.1
         </Text>
+        <Text style={styles.text} onPress={() => this.setUserTrackingMode(mapRef, this.userTrackingMode.follow)}>
+          Set userTrackingMode to follow
+        </Text>
         <Mapbox
           style={styles.map}
           direction={0}
@@ -128,16 +134,17 @@ var MapExample = React.createClass({
           showsUserLocation={true}
           ref={mapRef}
           accessToken={'your-mapbox.com-access-token'}
-          styleURL={'asset://styles/streets-v8.json'}
+          styleURL={this.mapStyles.emerald}
+          userTrackingMode={this.userTrackingMode.none}
           centerCoordinate={this.state.center}
-          userLocationVisible={true}
           zoomLevel={this.state.zoom}
           onRegionChange={this.onRegionChange}
           onRegionWillChange={this.onRegionWillChange}
           annotations={this.state.annotations}
           onOpenAnnotation={this.onOpenAnnotation}
           onRightAnnotationTapped={this.onRightAnnotationTapped}
-          onUpdateUserLocation={this.onUpdateUserLocation} />
+          onUpdateUserLocation={this.onUpdateUserLocation}
+          onLongPress={this.onLongPress} />
       </View>
     );
   }